Добавлен драйвер ISA DMA
Я ещё не проверял реально ли он работает, когда я сделаю драйвер FDC, SB16, или любого другого устройства которому нужно DMA, тогда проверю

+++ src/c/headers/kbc.h
@@ -10,7 +10,7 @@
 
 #define KBC_ACK 0xFA
 
-u8 kbc_read_data();
+u16 kbc_read_data();
 u8 kbc_init();
 void kbc_send_byte_port1(u8 byte);
 void kbc_send_byte_port2(u8 byte);
